@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace App\Application\Members;
 
-final class SignUpMember
+final readonly class SignUpMember
 {
     public function __construct(
-        public readonly string $firstName,
-        public readonly string $lastName,
-        public readonly string $emailAddress,
-        public readonly string $password
+        public string $firstName,
+        public string $lastName,
+        public string $emailAddress,
+        public string $password
     ) {
     }
 }
