@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace App\Domain\Members;
 
-final class MemberSignedUp
+final readonly class MemberSignedUp
 {
     public function __construct(
-        public readonly Id $id,
-        public readonly FirstName $firstName,
-        public readonly LastName $lastName,
-        public readonly EmailAddress $emailAddress,
-        public readonly StatusName $status
+        public Id $id,
+        public FirstName $firstName,
+        public LastName $lastName,
+        public EmailAddress $emailAddress,
+        public StatusName $status
     ) {
         //
     }
